refactor: rank 파일명 수정, css 주석 추가
