+++ src/css/index.css
@@ -127,7 +127,7 @@ section {
   flex-direction: column;
   align-items: flex-end;
 }
-
+/* align converter에 미리보기 기능을 추가한다. */
 #align-converter-container {
   display: flex;
   flex-direction: column;
@@ -143,7 +143,7 @@ label[for='align-converter']:hover::before {
   height: 200px;
 
   position: absolute;
-  right: -250px;
+  right: -220px;
 }
 
 #align-converter-container[data-visible-state='false']
